refactor(directory)
- 단순 컴포넌트와 그 외 App을 분리
